requirements.txt
streamlit
python-dotenv
langchain
langgraph
langchain-core
langchain-community
langchain-cohere
cohere
chromadb
tiktoken
pypdf
